// src/openapi/model/api-bank.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiBank
 */
export interface ApiBank {
    /**
     *
     * @type {number}
     * @memberof ApiBank
     */
    id?: number;
    /**
     *
     * @type {string}
     * @memberof ApiBank
     */
    bankCode: string;
    /**
     *
     * @type {string}
     * @memberof ApiBank
     */
    branchCode: string;
    /**
     *
     * @type {string}
     * @memberof ApiBank
     */
    nameKana: string;
    /**
     *
     * @type {string}
     * @memberof ApiBank
     */
    name: string;
    /**
     *
     * @type {number}
     * @memberof ApiBank
     */
    level: number;
    /**
     *
     * @type {string}
     * @memberof ApiBank
     */
    nameHalfWidth?: string;
}