// src/openapi/model/api-mypage-bank-account-to-add.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiMypageBankAccountToAdd
 */
export interface ApiMypageBankAccountToAdd {
    /**
     *
     * @type {number}
     * @memberof ApiMypageBankAccountToAdd
     */
    userId?: number;
    /**
     *
     * @type {string}
     * @memberof ApiMypageBankAccountToAdd
     */
    gmoBankName: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageBankAccountToAdd
     */
    gmoBankBranchName: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageBankAccountToAdd
     */
    gmoBankAccountType: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageBankAccountToAdd
     */
    gmoBankAccountNumber: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageBankAccountToAdd
     */
    gmoBankAccountName: string;
}